kiemtra01/mobile-service/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
